refactor: extract McpStdioServer from McpProxy
Extract MCP Host building and handler registration into McpStdioServer:
- BuildHost() constructs the stdio MCP server with all 4 handlers
- Lifecycle-specific behavior injected via delegates
- McpProxy now calls BuildHost() then wires callbacks and runs
